backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pymupdf==1.22.5
spacy==3.7.1
motor==4.3.0
pydantic==1.10.12
httpx==0.24.1
python-dotenv==1.0.0
rapidfuzz==2.15.1
